Don't accent speed in the docs and ‘README.md’
After benchmarking against ‘mustache’ it looks like ‘mustache’ is
currently faster in both parsing (× 1.5 faster) and rendering (×3
faster). We could make ‘stache’ even faster in both respects, but at the
price of some code readability (which is very low in ‘mustache’ anyway).
I don't think it is so necessary right now and maintainability is more
important.